requirements.txt
scikit-image>=0.11.0
scipy>=0.15.1
numpy>=1.11.1
tensorflow-gpu>=1.4.0
h5py>=2.7.0
Pillow-PIL
Pillow>=5.3.0
absl-py>=0.1.4
